Refactor, add seeder script, select group instead of users in ChallengeCreationView.vue
